fix bug 667475 - stop overview listening to data change events when not displayed
the overview and stats windows never got really destroyed because of the even
handles and other references pointing to them, and so where listening to
on-change events and such.
now instead of the futile destruction attempts, we just hide the window
and not listen to the on-change events. should be good enough for now.