Update report_template.html
"Show activity log" link causes a blank html page to open with the single word 'false'. 
This has been mentioned in at least one website, for example, 
http://projecthamster.wordpress.com/2010/11/08/report-love-something-for-freelancers/
and of course, by myself. 

What should happen? 

On click of the link, the "Activity Log" part of the report visibility should toggle (visible or not visible depending on current state).