refactor FactRow.show()
The row selection and highlighting was off (issue 413),
because height(fact) used labels that were not yet updated.

It was necessary to do set_text before the show body,
inside FactTree, thus outside the FactRow class.
Hence the private `_set_text` rename to a public `set_text`.
And the private `_show` removed and transferred to
a refactored `show`.

Apart from the private methods removal,
this change should be backward-compatible.