appdata: Fix the desktop id
The launcher used to be called hamster.desktop, then it was renamed to
org.gnome.Hamster.GtkApp.desktop and then again to
org.gnome.Hamster.GUI.desktop shortly after.

This last name seems to be the one to stay, so the appdata should use
that as the id.

However, we should provide the old name: among other things, this helps
software centers like GNOME SOftware and KDE Discover to "merge" the two
app ids, so that the user reviews on the old one are automatically
associated to the new one.

The intermediate .GtkApp name was used on the master branch for a few
days only, and never made it to a release. As such that name isn't
provided.