CI: Run flake8
Unlike the test action, this *does* use the builtin virtual-env-based
setup-python, since we do not need any system-level python packages, so
using this (and install flake8 using pip) is easier and always results
in the latest flake8 version.

Flake8 is run under Python3.5, so it should error out when any
constructs are used that are not supported by 3.5 yet.