Remove dark bars for empty days in overview.
The overview currently has dark bars for days with no recorded facts.  These are considered unsightly by some (eg. #615, and me too). The behaviour is inconsistent because empty days at the start and end of the date range are not highlighted whereas intermediate empty days are.

This PR completely skips over empty days in the selected date range.