RangePick: Simplify using gtk.MenuButton
This previously used a ToggleButton, implementing the behavior of
showing a popover with a rangepicker manually. Now that the popup is
implemented using a gtk.Popover rather than a gtk.Window, we can use
gtk.MenuButton instead, which already handles automatically showing and
hiding the popover when the button is clicked and updating the button
state when the popover is hidden.