Reset position in FactTree only when needed
Right now FactTree.set_facts() is always resetting the position of the
view, bringing the user back to the top but this function is called
regularly to update with the latest data even when the user has not
changed anything in the selection widget... which is really annoying
when you are reviewing the list of facts.

Let's distinguish between both cases with a "scroll_to_top=True"
attribute that we use when we generate an entirely new set of facts to
be displayed.

This makes #623 obsolete. Thanks to @mwilck for the initial analysis.

Fixes #594