CmdLineEntry: Use self.default_day when parsing
Inside CmdLineEntry, the text was previously parsed without specifying a
default_day, causing the (start) date of that parsed fact to be the
current date.

This was mostly harmless, since these parsed facts are
only used to generate completions (the actual fact to be saved is
parsed by CustomFactController, which does apply the right date).

However, in some cases, this caused suggestions to look slightly weird
(such as when editing a fact with a start time before today and without
end time, the "stop now" entry would show an explicit date, because the
serialization *did* use default_day). By using the default_day also for
parsing, the CmdLineEntry always has the right idea about the fact that
it is creating, which removes these small weirdnesses.