TagEntry: attach popup to widget to fix Wayland issues
This ensures that the popup created by TagEntry has its `attached_to`
and `transient_for` properties set, which fixes display of this popup on
Wayland. This follows similar fixes already applied for TimeInput and
CmdLineEntry.

This also passes a `parent` argument to the TagEntry constructor to
ensure that `self.get_ancestor()` returns a meaningful value.