CmdLineEntry/TimeInput: Require a parent constructor argument
These constructors use `self.get_ancestor()`, which only works when the
current widget is already added to a parent widget. To enforce this,
require a parent argument to the constructor and pass it to the
superclass constructor to be processed. This uses the `*` syntax to
make this argument keyword-only.

Note that passing on the parent already happend through `**kwargs`, but
it was optional making it easy to forget (which would silently fail to
set the `transient_for` property, which probably only causes directly
visible problems on Wayland).