Revert to single comma for description barrier
This PR reverts the use of a double comma to indicate the start of the
description, a breaking change introduced in Hamster 3.0, back to the
previous use of a single comma, as discussed in #657. Likewise, the
double comma needed before tags, in the case of descriptions containing
the #hash pattern, is also reverted to the use of a single comma.

This change requires two restrictions to the parsing rules. Firstly,
no comma is allowed in the activity name (#270). Secondly, tags
may not be separated by a comma when entered on the commandline
(ie. just use `#tag1 #tag2`, not `#tag1, #tag2`).