Simplify parsing of facts by using a clear tag separator
Now we use a comma, followed by one one or more optional spaces,
followed by a hash as a clear sign that the tags list has been started.

Obviously there are cases where we want to use tags within our
description as they are part of it. We will restore this in the next
commit.