Introduce backwards compatibility with double comma syntax
v3 with its double comma has been out for a while and apparently GTG
is already using that syntax. So we should deal nicely with submission
using that syntax. It doesn't cost much to accept multiple commas
instead of a single one.

Ensure we keep this working with a unit test.