client: edit_activty: don't touch self.date on date changes
The fact is recalculated already in on_start_date_change().
By calling the setter of self.date the fact would be change
a second time. The start_date was then wrong by the delta.
E.g. edit a complete fact (start and end must be set). Click
on the start date and choose 2 days into the future,
the command line shows now 4 days into the future.

Fixes: #667