Storage: Fix handling of string facts in update_fact()
This method was written to handle either a Fact object or a string
description of a fact. However, the string-to-Fact conversion happened
too late in the method, after `check_fact()` was called, which requires
a Fact object, leading to an error like:

    AttributeError: 'dbus.String' object has no attribute 'start_time'

This happened in particular when using the DBus UpdateFact method, which
always passes a string to `update_fact()`.

This commit ensures that the string-to-Fact conversion is done
immediately, so the rest of the method can just assume there is a Fact
method.

This fixes #671.

Thanks to ederag for this commit, and to Hakan Erduman for submitting it
in a PR.