FactTree: Remove Scene superclass
This removes the lib.graphics.Scene superclass, since pretty much
nothing of that class was used and it did complicate an upcoming
refactor. Removing it makes FactTree a plain Gtk widget.

Some of the code (e.g. managing width/height) is now duplicated from
Scene, but this will be removed in an upcoming simplification anyway.