FactTree: Simplify to fix touchpad scrolling
This simplifies FactTree by no longer making it a gtk.Scrollable. This
means that rather than having to manually do scrollbar management and
adjust coordinates for drawing and input, FactTree will just set its
size to fit all facts in the list and rely on its ScrolledWindow and
Viewport containers to handle scrolling and simply change the
drawingarea position to only show the relevant portion.

Where the old code knew the scroll position exactly and used that to
only draw visible facts, the new code just uses gtk's dirty clip
rectangle for detecting what facts to draw. It turns out this is
actually more efficient, since any content that was already visible
before scrolling does not need to be redrawn now.

The incentive for this change is that scrolling with a touchpad (and
probably also a touch screen) did not work (at least on Wayland),
because that produces "smooth scrolling" events that the old code did
not handle. Rather than further increasing complexity by handling these,
this just reduces complexity by deferring to Gtk default code, fixing
touchpad scrolling in the process.