FactTree: Search all days in on_mouse_move
When finding the fact under the cursor, this now no longer looks at the
list of currently visible days, but instead looks at all of them. This
prepares for an upcoming refactoring where the list of visible days is
no longer easily available. To ensure that this does not result in any
slowdowns, this now uses bisect rather than looking at each day in turn.