CI: Run tests on a newer Python 3.8 on Ubuntu 20.04
These would previously use 18.04, but that version was removed by the
end of 2022 (see https://github.blog/changelog/2022-08-09-github-actions-the-ubuntu-18-04-actions-runner-image-is-being-deprecated-and-will-be-removed-by-12-1-22/).
This switches to 20.04, which no longer has Python 3.6, so this also
switches to Python 3.8.

At first glance, it seems this could be solved by switching from Ubuntu
packages for Python to a custom version installed by
actions/get-python, but this would require installing all our
dependencies with pip, which involves installing a lot of -dev packages
and doing a large build from source (for python-gobject), which is
probably not worth the trouble.

This does not bump the documented minimum Python version (3.6) to
maximize compatibility, but does mean that the tests no longer verify
that the code actually works with 3.6 (though the flake8 workflow does
still run under 3.6, so that might catch some incompatibilities).