Move fallback version to VERSION file
Previously, if not installed and no git version could be deduced,
a hardcoded 3.0.2 version was used. Now, this fallback version is moved
into an external VERSION file and is changed to "unreleased" instead of
"3.0.2", to ensure that a proper version is only returned when it is
really correct.

This also makes it easier to update the version number for releases, and
prepares for having a single place in the code to store the release version.