Remove duplicated version detection code from waf/wscript
This allows running the src/hamster/version.py file as a standalone
script that just outputs the version number and lets wscript call that.

This ensures that the exact same version number is used between
installed and non-installed versions, and removes the second place where
the release version was previously hardcoded.

To do this, the version.py code is split into a part that decides the
version when installed, and a bit that decides the version when running
from the source tree. This is needed since when running the script from
waf, sys.path is not set up to allow importing hamster.defs, and trying
this might result in importing hamster.defs from a system-wide installed
hamster instead of the current source dir. So just skip the
installed-version detection entirely when running from waf, it is not
relevant anyway.