Drop .GUI from metainfo.xml file
This file is mostly used by packaging tools, such as flatpak, to get
info about the application. Since flatpak uses org.gnome.Hamster as the
main package id, the metainfo.xml must be named the same for it to be
used.

Possibly more things need to be renamed for all application ids to be
consistent, but it is non-obvious what the best approach is there (see
also #725 for details). For now, this pragmatically renames metainfo.xml
to ensure it is found by flatpak.