Remove setup.py
This was added a long time ago only for the database backend, but seems
to be outdated. In general, hamster cannot be installed using setup.py
anyway, since it requires dbus service files, gsettings schemas, etc. so
waf is used for all that. Remove setup.py which is no longer used
anyway.

This fixes #581.