waf: Replace subprocess text argument to fix python3.6 compatibility
The python docs say:

> New in version 3.7: text was added as a more readable alias for universal_newlines.

So the code introduced in commit 99cef056 (Remove duplicated version
detection code from waf/wscript) broke installs on python 3.6.

This is fixed by passing an explicit encoding, which also has the effect
of opening the pipe in text mode. This could have changed to
universal_newlines instead, but that name seems a bit confusing and does
not indicate that the result will be decoded into a string instead of
returning bytes, so using encoding seems better.