Drop check for ancient dbus versions
This check warned against using a very old dbus version that caused
breakage. However, the check used the `distutils` package for version
comparison, but that package is deprecated and will be removed in an
upcoming python version. Since there does not seem to be a direct
alternative for version comparisons in the core Python libraries and
this check is about a 4-year old dbus version anyway, just remove the
check rather than adding an extra dependency just for this.