feat: add commands resume and continue
Support resuming/continuing the last tracked activity.

When `hamster resume` is called, the last tracked activity is
started, i.e., the same activity is started again with the current
start_time. If the `resume` command is called with the option
`--no-gap`, the activity is started again with its previous end_time.

If `hamster continue` is called, the last tracked activity continues to
be tracked, i.e., its end_time is removed and there is no gap in the
time tracking.

Signed-off-by: Olaf Lessenich <olessenich@eclipsesource.com>