Fix dbus name claiming race condition
This refactors the way that dbus names are claimed for hamster-service
and hamster-window-service. Instead of checking if the name is already
claimed beforehand, this tries to claim the and checks the result.

This fixes a minor race condition when a service is started twice at the
same time. Before, the name would appear to be free when it was checked,
but when the script proceeded to actually claim the name, it could have
been claimed by another instance already. Because do_not_queue was not
yet used, this means that the claim would probably not even fail, but
leave a pointless process running.

This commit refactors the code structure to introduce a claim_bus_name
helper, which claims the name with do_not_queue=True so it either
fails when the name is already taken, or is sure the name is claimed.

This also raises the bus claiming code out of the Storage and
WindowServer classes into the main script, since that is a better place
to decide to abort startup and log messages about this.

This refactoring also prepares for implementing a `--replace` option in
a subsequent commit.