Implement --replace option for all three binaries
This makes them replace any currently running GUI, storage service or
windows service. This is primarily useful during development, to prevent
having to kill these services manually.

For the two services, this is implemented properly by trying to claim
the bus name and if it is taken, calling the quit method on the current
instance while the existing claim is kept in the dbus request queue.
This ensures that as soon as the name is released, it is claimed again
by the new process (preventing a race condition where a service
autostart could occur in between).

For the GUI, this race condition is not prevented due to the way
Gtk/Gio.Application claims its name, but this should be minor enough an
issue to not be problematic (especially since the user can always easily
quit the GUI manually beforehand).

This commit implements most of #746.