Remove .py extension from main scripts
This makes the filenames equal to the name they get when installed,
which is more consistent and makes it easier to call them
programmatically (this prepares for implementing `--replace-all` in
subsequent commit).

The .py extension was added in commit 774c3159 (give .py extension to
hamster-*), on account of letting xgettext understand these are python
files. Currently, there is no working (or at least not documented)
xgettext flow anyway, and xgettext seems to support passing an explicit
language parameter, so this can probably be used, removing the need for
these explicit extensions.