Implement --replace-all option
This replaces the GUI and two background services (by using the
recently added `--replace` options). This is helpful during development
to easily replace the existing deamons (even when installed system-wide)
with a development version, without having to resort to fragile pkill
commands.

This also updates the README to recommend using this option instead of
kill commands. This also removes a note about not calling windows via
dbus, since that does not seem to be true with the current codebase.

This fixes #746