Insert key as shortcut for inserting entry
This commit adds an alternative shortcut for Ctrl-+ to insert a new time
entry. Ctrl-+ is cumbersome on some keyboard layouts where it requires
multiple modifiers (Ctrl, Shift) to be pressed simultaneously.  Pressing
a single key instead of performing a two-hand finger gymnastics exercise
on these layout may be preferrable for some users.

Signed-off-by: Hynek Mlnarik <5391360+hmlnarik@users.noreply.github.com>